final v

+++ src/main/java/ma/ensat/test_struts/actions/FiliereAction.java
@@ -5,13 +5,16 @@ import com.opensymphony.xwork2.ActionSupport;
 import com.opensymphony.xwork2.ModelDriven;
 import com.opensymphony.xwork2.interceptor.annotations.Before;
 import com.opensymphony.xwork2.interceptor.annotations.BeforeResult;
+import ma.ensat.test_struts.Services.EleveService;
 import ma.ensat.test_struts.Services.FiliereService;
 import ma.ensat.test_struts.Services.FiliereServiceImp;
+import ma.ensat.test_struts.models.Eleve;
 import ma.ensat.test_struts.models.Filiere;
 import org.apache.struts2.convention.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Controller
@@ -19,10 +22,14 @@ public class FiliereAction extends ActionSupport implements ModelDriven {
     private Filiere filiere;
     private List<Filiere> filieres;
     private String codeId;
+    private List<Eleve> listeEleve= new ArrayList<>();
 
     @Autowired
 
     private FiliereService service;
+
+    @Autowired
+    private EleveService eleveService;
 
     @Action(value = "allFilieres", results = {
             @Result(name = "success", location = "/afficheher_Filiere.jsp"),
@@ -31,6 +38,8 @@ public class FiliereAction extends ActionSupport implements ModelDriven {
 
     public String getAll() {
         filieres = service.getAll();
+
+
         return "success";
     }
     @Action(value="loadFiliere")
@@ -40,6 +49,8 @@ public class FiliereAction extends ActionSupport implements ModelDriven {
         System.out.println("the size of filieres is: "+ filieres.size());
 
     }
+
+
 
 
     @Action(value = "addFiliere", results = {
@@ -58,6 +69,7 @@ public class FiliereAction extends ActionSupport implements ModelDriven {
     })
     public String editFiliere() {
         filiere = service.getCode(codeId);
+
         return "success";
     }
 
@@ -77,8 +89,20 @@ public class FiliereAction extends ActionSupport implements ModelDriven {
             @Result(name = "error", location = "/error.jsp")
     })
     public String deleteFiliere() {
-        service.delete(codeId);
-        return "success";
+        if (codeId != null && !codeId.isBlank()) {
+
+            filiere = service.getCode(codeId);
+            listeEleve = service.getAllEleves(filiere);
+            for (Eleve eleve : listeEleve) {
+                eleve.setRef_fil(null);
+                eleveService.update(eleve);
+            }
+            service.delete(codeId);
+            return SUCCESS;
+        }
+        return ERROR;
+        //service.delete(codeId);
+        //return "success";
     }
 
     public Filiere getFiliere() {
